Build and copy changes
