+++ GlockTheClock/assets/scripts/Client/EventsList.lua
@@ -1,0 +1,3 @@
+--New events are to be required here for loading
+
+--require('sampleEvent')
