+++ GlockTheClock/assets/scripts/Client/PrefabsList.lua
@@ -1,0 +1,7 @@
+--New Prefabs are to be required here for loading
+
+require('Bullet')
+require('Enemy')
+require('sampleScene-config')
+require('Platform')
+require('Player')
