+++ GlockTheClock/assets/scripts/Client/main.lua
@@ -18,12 +18,14 @@ end
 
 -- setWindowVisible("TitleWindow", true)
 
--- createButton("PushButton", "CEGUI es facilito", "TaharezLook/Button",vec2:new(300, 200), vec2:new(200, 50))
+-- createButton("PushButton", "CEGUI es facilito", "TaharezLook/Button", "DejaVuSans-24",vec2:new(300, 200), vec2:new(200, 50))
 -- setButtonFunction("PushButton","showTaiFighterUI")
 
--- createButton("ExitButton", "Exit", "TaharezLook/Button",vec2:new(300, 300), vec2:new(200, 50))
+-- createButton("ExitButton", "Exit", "TaharezLook/Button", "DejaVuSans-12",vec2:new(300, 300), vec2:new(200, 50))
 -- setButtonFunction("ExitButton","exitCallback")
 
 -- showTaiFighterUI()
+
+-- setProgressBarValue("DimensionBar", 0.2);
 
 LOG("main.lua completed")
